Log: scheduler: correct lifetime deprecation warning
Contrary to the log message as it was, rules cannot simply be moved from the
lifetime element to its parent constraint element. Colocation, ordering, and
ticket constraints do not support rules at all, and location constraints
support them with different semantics.

The feature is undocumented and has been deprecated since 1.0.8, so it should
be fine to not recommend any alternative.