Test: cts-lab: Drop -t argument from cts-log-watcher
It was introduced in commit 1d521fc5 and has never done anything.
cts-log-watcher skips the -t flag and the argument that follows it. The
commit message seems unrelated.

It looks like -t is a remnant of a draft implementation, and that it was
abandoned in favor of "-p CTSwatcher". It likely would have been used
for filtering log messages based on a tag set by something like the
logger command's -t option.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>