Low: tools: use better value for crm_resource --force-* timeout
This drops the pe_get_configured_timeout() function, because it duplicated what
other functions already do, but poorly. It did not handle rsc_expression and
op_expression in rules, did not default to the minimum-interval monitor timeout
for probes, etc.

Also, use guint for the timeout, since that's how it's eventually used.