API: schemas: Change the type of a ticket's id attribute.
It's possible to have multiple tickets with the same ID in the <state>
section.  Normally this is not a problem because there's no schema for
this section.  However, when these tickets become part of an XML blob
returned from the public API, they are now subject to a schema.
Duplicate IDs are not allowed, so the result will not validate.