Doc: cts-lab: Document log aggregation detection
Support has always been minimal (can't specify a log file or use
journald log forwarding) and undocumented, but keep for backward
compatibility. It may be worth expanding in the future. I would expect
cts-lab to be more efficient with log aggregation set up, since it
eliminates the SSH calls for log watching. The logs are pushed as
they're generated, instead of pulled on demand.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>