Fix: libcrmcommon: Revert daemon metadata output to pre-3e84f934
We've recently deprecated daemon metadata commands in favor of
crm_attribute --list-options=cluster, using the "option-list" message
formatter function. Along the way, we've made a number of changes to
clean up and standardize the output.

Unfortunately, some external tools (for example, pcs and crmsh) depend
on very specific strings in the old output format. For example, pcs
ignores the <content> element containing the list of possible values;
instead it relies on the redundant "  Allowed values:" string in the
<longdesc>. It also validates the output against its own custom version
of the OCF resource agent schema, which does not allow the "advanced"
and "generated" attributes.

crmsh has some similar dependencies, though I'm not aware of the
details.

These tools should eventually be updated to use crm_attribute and the
new output format. However, in the meantime, the daemon metadata output
should go back to the old format, so that the tools don't break upon a
Pacemaker upgrade.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>