Build: Promote gnutls to a required build dependency.
This is necessary because we are going to use it for its hashing
functions.  With this done, a lot of the complication in configure.ac
can also go away because we can just require the header file and check
pkgconfig for the library.

And then with that done, there's no real advantage to having
BUILD_REMOTE anymore - we always have gnutls, so we don't need to
conditionalize that build feature.