Refactor: libcrmcommon: Use gnutls for hashing.
This is just about as easy as what we were already getting from gnulib.
Plus, it's a shared library that just about everyone already had, with
its own upstream that we don't have to bundle and update in our source.

My lame unit test passes, as do all the regression and scheduler tests.