Refactor: libcrmcommon: Drop utf8_bytes()
A lot of the complexity of utf8_bytes() was for dealing with the fact
that the C standard doesn't specify the size of a byte. UTF-8 characters
come in 8-bit chunks. utf8_bytes() detected how many 8-bit bytes were in
a UTF-8 character and then converted that to a number of C bytes.

As it turns out, GLib assumes an 8-bit char and is not supported on
platforms that use a wider char [1]. Also, POSIX mandates in Definition
3.84 [2] that a byte is "exactly an octet" (8 bits). (Note, however,
that POSIX-compliance is not guaranteed on Linux or FreeBSD.)

Since we already depend so heavily on GLib, it seems reasonable to
assume an 8-bit byte. Then we can use g_utf8_next_char() to get a
pointer to the next UTF-8 character in a string. Determining the number
of bytes to skip is implemented more efficiently there (by indexing into
an array), and this avoids reinventing the wheel and adding clutter.

Note: the GLib docs recommend calling g_utf8_validate() on the string
before calling g_utf8_next_char(). However, all GLib functions assume
that strings are encoded as UTF-8 [3]. I suspect most of Pacemaker and
parts of libxml2 would fall apart if we received non-UTF-8 strings. Our
escape-XML functions don't seem like the place to start validating the
encoding.

[1] https://discourse.gnome.org/t/glib-utf-8-functions-assume-8-bit-guchar/20224/2
[2] https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_87
[3] https://docs.gtk.org/glib/programming.html#utf-8-and-string-encoding

Closes T801

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>