API: libcrmcommon: Deprecate fix_plus_plus_recursive()
This function does not work correctly -- it doesn't have access to
the old value, which it needs in order to correctly expand the new
value. There are more details and a demonstration in the previous commit
message.

It was added in a4b42975 for updating fail counts. Now,
update_failcount() sends its update to attrd, which handles it correctly
via attrd_expand_value().

There are no more internal callers of fix_plus_plus_recursive(), and
it's likely been broken for years, so we deprecate it here.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>