Low: libpacemaker: pcmk__inject_failcount should set an integer value
Currently it sets the new failcount value to "value++". The idea was
that this should increment the existing value. In practice, it always
sets the value to 1. This is due to a bug in fix_plus_plus_recursive
that will be fixed in an upcoming commit. To summarize,
fix_plus_plus_recursive() doesn't have access to the initial value, so
it treats the initial value as zero when incrementing it.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>