Refactor: libcib: Drop dead code for PCMK__XA_REPLACE
The dropped section's purpose was to delete certain items from the
target XML so that they would be replaced by the next step of
update_cib_object().

However, target is always a brand new, empty XML node. So this block can
never match anything.

The only part we need to keep is the removal of the PCMK__XA_REPLACE
attribute if it exists, to preserve backward compatibility since it has
had special meaning up to now. When we can break behavioral backward
compatibility, we can drop this and drop the PCMK__XA_REPLACE
definition.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>