Low: scheduler: compare nvpair block IDs case-sensitively
Also, refactor to get the ID when needed instead of putting it in sorted_set_t
ahead of time. This will be trivially more efficient with the common case of a
single name/value pair block, and trivially less efficient as the number of
blocks increases. Real-world use cases typically only have 1 to 3.