Test: cts-cli: Update for marking dirty in pcmk__xml_mark_created()
xml_create_patchset() checks whether the PCMK_XE_CONFIGURATION element
is dirty. If so, it bumps the epoch and resets num_updates. If not, it
bumps num_updates.

Before the previous commit, we missed an epoch bump because
pcmk__xml_mark_created() did not guarantee that all parents would be
marked dirty.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>