Feature: CIB: deprecate support for unknown schemas in validate-with
As far as I can tell, both the live cluster and CIB_file will refuse to use a
CIB with an unknown schema, so we should be able to just drop support, but for
extra safety we'll just deprecate it now.

THe pacemaker-0.6, pacemaker-0.7, pacemaker-1.1, and transitional-0.6 schema
names are allowed by xml/versions.rng but have no actual schema, which means
they are treated as unknown and rejected. Again for extra safety, we'll wait to
remove these from versions.rng.