Low: libcrmcommon: max should apply even if transform is false in update_validation()
Previously, update_validation()'s max argument (specifying the maximum schema
index to update to) was respected only if the transform argument were true.

Now, respect it even when transform is false. No internal callers ever passed a
nonzero max with a false transform, so it is irrelevant internally, but it is
in line with expectations and is easier to follow logically.

Technically this is a change in behavior, but especially considering this
function will be deprecated as part of the public API and has no known external
callers, I'm considering it a bug fix.