Low: lrmd: Warn on insecure and undocumented key handling behaviors.
If PCMK_authkey_location is given in /etc/sysconfig/pacemaker, but the
key doesn't exist or is unreadable, we will fall back to trying to read
the key from default locations.

If /etc/pacemaker/authkey doesn't exist, we will fall back to trying to
use /etc/corosync/authkey.

Both of these are undocumented behaviors that end up loading a key from
a location the administrator does not expect.  Mark them as deprecated
so we can remove them in 3.0.0.

Ref T304