Fix: libcrmcommon: don't assume next schema will validate when not transforming
Previously, if the XML had a configured schema, update_validation() reasonably
assumed the XML validated against it and started checking with the following
schema.

However, it also set local_best to the following schema, which was a bug if the
following schema didn't validate. That bug accumulated from numerous unrelated
changes over the years.

Now, just drop the assumption and always start with validating the original
schema, which avoid the bug and simplifies the function.