Low: libcib: Fix handling of return values in CIB attr update/delete
Fixes regression introduced by 3d16e088. That commit message says:

"This function doesn't return a standard pacemaker return code.  It
returns a negative value on error and the positive call ID on success."

But these may be synchronous calls, in which case the return value is
the (legacy) Pacemaker return code, not the call ID.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>