Refactor: libcrmcluster: New pcmk__cluster_forget_cluster_node()
To replace reap_crm_member().

Use a separate function body from the soon-to-be-deprecated
reap_crm_member(), because there's no need to return the number of
matches.

Also use "remove" terminology rather than "purge". This is debatable as
"purge" is commonly used in caching contexts. My thought is that
"remove" may be more familiar to nonnative speakers and may feel less
drastic.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>