Feature: schemas: Add a schema for the new crm_resource output.
Unfortunately, we require another separate schema for resource attribute
stuff.  The existing schemas in crm_attribute and the CIB are too
permissive for what we're going to output here - just the resource ID,
attribute set ID, and the attribute itself.  There's not a great way to
reuse just those bits from elsewhere.