Refactor: libcrmcluster: Don't support PID as cluster message type
It's not type-correct (a PID isn't generally an enum value), and it's
not useful. The message sender and destination types are used only for
logging via msg_type2text(), and the sender's PID is logged
unconditionally.

If pcmk__cluster_parse_msg_type() is unable to parse a real message type
from a string, return crm_msg_none. Don't try to treat the text as an
int.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>