Refactor: libcrmcluster, controller: Drop crm_msg_ais internally
Except for the to/from string functions.

Previously, it was used only trivially as an initial value in the
controller. send_cpg_text() asserts if the message type is crm_msg_ais,
and there are no other currently supported cluster layers, so it makes
more sense to use crm_msg_none as the initializer.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>