Refactor: libcrmcommon: Drop pcmk__tracking_xml_changes()
Most of these changes are straightforward. A slight exception is the
"lazy" case. When xml_diff_old_attrs() is called, pcmk__xf_tracking is
always set (via xml_calculate_changes()). So

!pcmk__tracking_xml_changes(new_xml, TRUE)

is simply checking that the lazy flag is NOT set before proceeding to
mark the attribute moved.

It's probably worth renaming pcmk__xf_lazy to be more clear but that can
come later.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>