Refactor: libcrmcommon: New pcmk__xml_patchset_versions()
Preserve the body of xml_patch_versions() because the new function
returns EINVAL in the edge case where we fail to parse version numbers
as integers.

The order of arguments changes. I think it's clearer to use "source" and
then "target". In v1 patchsets, we used "add"/"remove", but we can start
moving away from that.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>