Refactor: libcrmcommon: New pcmk__xml_patchset_add_digest()
Preserve body of patchset_process_digest() because the new function
returns without adding a digest if the target is dirty.

Drop with_digest argument, and instead check in the callers. There are
only two callers, so it's not terribly clunky. It felt awkward to take
an argument for "should we add a digest?" in a function whose purpose is
to add a digest.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>