Feature: tools: Convert iso8601 to formatted output.
This adds XML support as well, which is somewhat pointless since this
program doesn't really see a lot of use, but it's consistent with the
rest of our command line tools.

Additionally, here the XML output uses text nodes instead of putting
everything into attributes.  This is because the tool can potentially
output multiple messages, one after the other.  We never have all the
data we need at once, so using text nodes is a cleaner way to structure
the code.

Fixes T109