Refactor: CTS: Add a docstring for the main function.
Note that another way to do this would be to get rid of the main
function and instead put everything under the "if __name__ == "
conditional.  However, doing that makes all the variables global, which
then means they need to obey different rules and we also need to make
sure that functions don't define variables with the same names.

That's a lot more work for relatively little benefit, so just add the
docstring instead.