Refactor: cts: Use test objects in cts-scheduler.
Instead of lists that typically have two elements but sometimes have
three, use a SchedulerTest object.  And then add a SchedulerTestGroup
object that for the moment just contains lists of tests.

Not much else is happening with these for now, but in the future it's
possible these objects could grow other functionality.  For the moment,
their main purpose is clarity and getting rid of a ton of pyflake
indentation warnings.