API: fencer, libstonithd: Drop support for "action" fencing param
Previously, if "action" was specified as a fencing resource parameter,
it acted as a deprecated alias for pcmk_<action>_action. Now we ignore
its value if present.

We continue to set the parameter internally, based on the action that
we're performing, before passing the option list to the fence agent. So
the fence agent can still check which action is occurring.

This breaks API backward compatibility.

Closes T773

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>