Doc: libpe_status: Drop COMPAT note in generate_fail_regex()
We no longer support rolling upgrades from < 2.0.0. There are regression
test input CIBs from old Pacemaker versions where an anonymous clone has
a fail count with an instance number. However, this is indistinguishable
from a previously unique clone that failed and then was converted to
anonymous. We keep those failcounts with instance numbers in the test
inputs, in order to test that scenario.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>