Refactor: python: Copy the body of _new_cmd into add_cmd.
The add_cmd function is going to become the one stop shop for adding
test commands, replacing all the other add_cmd_* functions.  There's
really no need for them anymore now that everything is a keyword
argument.

The first step is simply copying the body of the function over, updating
docs and keyword args.  I don't believe any callers need to be changed
for this.