Low: tools: Don't skip formatting if running crm_simulate interactively.
Somewhere along the line, we changed the interactive behavior of
crm_simulate to not enable fancy text formatting.  This is necessary for
non-interactive use (like when running cts-scheduler), but for
interactive use, it is hard to read.

Change the conditional on when fancy text is enabled to be always, as
long as we are not showing scores in quiet mode.  This seems to be the
combination that differentiates interactive use from regression test
use.

Fixes T833