Fix: scheduler: consider group locations when member is explicit colocation dependent
When location constraints are applied for a group, the location constraint
score is added to the first member's allowed nodes, but not subsequent
members'. This avoid counting the scores multiple times due to the internal
group colocations.

However that means that if a group member other than the first is the explicit
dependent in a user-configured colocation, the colocation primary will not
consider the group's location preferences.

Special-case this situation by backing up the member's allowed nodes, adding
the group location preferences to it before determining the best node score for
colocation purposes, then restore the original allowed nodes. It's clunky, but
the least intrusive approach I could come up with.