Log: scheduler: drop incorrect error logs in pcmk__colocation_affects()
For example, if a bundle with a primitive has a mandatory colocation with
another resource, pcmk__colocation_affects() will be called for an instance of
the bundle clone as the dependent. The instance runs on the bundle node, which
is not the same as the primary resource's node, but the colocation is satisfied
as long as the bundle container is running on the primary resource's node.

Rather than special-case bundles, just drop the log messages, since they don't
add anything useful that isn't already covered elsewhere.