Low: libpacemaker: Apply promotion priority to cloned group instance
Rather than only to its first member. We sort instances for promotion
based on each instance's priority. In the case of a cloned group, each
instance is a group, and we call pcmk__apply_coloc_to_priority() for the
group's first member. The first member's priority is not visible to the
instance sorting code.

This doesn't change the scheduled actions in any existing regression
tests, but it does change a few scores and may prevent some bugs in edge
cases.

Continue assigning priority to the first member too, in case of any
effects on action priority. (priority="1000000" disappeared from the
transition graph of one regression test when I tried assigning to only
the group, although overall behavior did not change.)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>