Test: cts-scheduler: Test for doubled colocation scores for promotion
In task T832, a user reported that colocation scores were sometimes
being applied twice when determining which clone instance to promote.
So far, this has been reproducible only when the primary resource in the
colocation will be scheduled to move due to a mandatory location
constraint. There may be other possible triggers; however, that's the
one we use in these tests.

Currently, coloc-optional-promoted-dependent-stays-1 and
coloc-optional-promoted-dependent-moves-2 tests exhibit the correct
behavior, and the other two exhibit incorrect behavior. However, even in
the tests that behave correctly, the scores are incorrect.

Ref T832

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>