Test: cts-scheduler: Update outputs for double colocation fix
Scores that change in the bug-lf-2358 and use-after-free-merge tests do
so correctly.

In the coloc-cloned-group-promoted-dependent2 scores, we see that the
stickiness is no longer added to the promotion priority. This is okay,
and we don't need to consider stickiness in order to break ties. The
instances are numbered/ordered starting with the currently promoted
ones, which avoids unnecessary demotion/promotion.

Ref T832

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>