Refactor: controller: Get node UUID directly instead of from cluster obj
This way we can drop the uuid member from pcmk_cluster_t. It's relevant
to a node, not to the cluster itself.

We can make an argument for storing some local node data in the cluster
object, but that seems hacky.

(This is not a true UUID but renaming is a pain right now.)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>