API: libcrmcluster: Drop crm_peer_seq global variable
It's used only by the controller, so replace it with
controld_globals.peer_seq. As noted in the TODO comment, it's unclear
exactly what the relationships are among this, highest_seq, and
controld_globals.membership_id. Sorting that out is out of scope of this
commit.

This breaks API backward compatibility.

Ref T793

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>