Med: libcrmcommon: Use glib for calculating MD5 checksums.
We are still experiencing FIPS problems with gnutls, even after the
previous fix.  So, instead just switch to using glib for calculating
checksums.

glib doesn't appear to be FIPS aware.  Additionally, it has separate
code for checksums and HMACs.  If it starts caring about FIPS in the
future, it's likely that it will only be for the HMAC code.  FIPS
appears to only care about crypto related code, and glib seems to take
the position that checksums are not crypto.

The most significant downside here is that error reporting is lacking.
If it fails, we can't really tell why.