Refactor: CTS: Split crm_resource test out of tools.
This one is a little complicated.  The crm_resource tests are so tied up
in the middle of a big line of other tools tests that they operated on a
fairly complicated CIB.  Now, they start with almost nothing and we only
add things to the CIB that are required for the test.  This makes it a
little harder to confirm that the regression output has not changed.