Feature: xml: Allow previously disallowed cluster properties
These properties were disallowed beginning with schema version 3.0, to
avoid confusion.

As of the 2.x schemas, they were allowed and affected Pacemaker
behavior. Support for these properties was dropped when we introduced
the 3.0 schemas. We disallowed these properties in 3.x schemas so that
users would not mistakenly configure them expecting to affect behavior.

Keep the unsupported nvpair section, because we'll soon add to it.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>